chore: nest swapper-specific types
