+++ packages/swapper/src/manager/SwapperManager.test.ts
@@ -4,10 +4,11 @@ import { KnownChainIds } from '@shapeshiftoss/types'
 import Web3 from 'web3'
 
 import { SwapperType } from '../api'
-import { ThorchainSwapper, ZrxSwapper, ZrxSwapperDeps } from '../swappers'
-import { CowSwapper, CowSwapperDeps } from '../swappers/cow/CowSwapper'
+import { CowSwapper, ThorchainSwapper, ZrxSwapper } from '../swappers'
+import { CowSwapperDeps } from '../swappers/cow/CowSwapper'
 import { ThorchainSwapperDeps } from '../swappers/thorchain/types'
 import { WETH } from '../swappers/utils/test-data/assets'
+import { ZrxSwapperDeps } from '../swappers/zrx/types'
 import { SwapperManager } from './SwapperManager'
 
 describe('SwapperManager', () => {
