+++ packages/swapper/src/swappers/zrx/getZrxTradeQuote/getZrxTradeQuote.ts
@@ -6,11 +6,10 @@ import { bn, bnOrZero } from '../../utils/bignumber'
 import { APPROVAL_GAS_LIMIT } from '../../utils/constants'
 import { normalizeAmount } from '../../utils/helpers/helpers'
 import { getZrxMinMax } from '../getZrxMinMax/getZrxMinMax'
-import { ZrxPriceResponse } from '../types'
+import { ZrxPriceResponse, ZrxSupportedChainIds } from '../types'
 import { DEFAULT_SOURCE } from '../utils/constants'
 import { baseUrlFromChainId } from '../utils/helpers/helpers'
 import { zrxServiceFactory } from '../utils/zrxService'
-import { ZrxSupportedChainIds } from '../ZrxSwapper'
 
 export async function getZrxTradeQuote(
   input: GetTradeQuoteInput
