+++ packages/swapper/src/swappers/zrx/types.ts
@@ -1,4 +1,8 @@
-import { SwapSource } from '../../api'
+import { avalanche, ethereum } from '@shapeshiftoss/chain-adapters'
+import { KnownChainIds } from '@shapeshiftoss/types'
+import type Web3 from 'web3'
+
+import { SwapSource, Trade } from '../../api'
 
 export type ZrxCommonResponse = {
   price: string
@@ -18,3 +22,14 @@ export type ZrxQuoteResponse = ZrxCommonResponse & {
   data: string
   gas: string
 }
+
+export interface ZrxTrade extends Trade<ZrxSupportedChainIds> {
+  txData: string
+  depositAddress: string
+}
+
+export type ZrxSwapperDeps = {
+  adapter: ethereum.ChainAdapter | avalanche.ChainAdapter
+  web3: Web3
+}
+export type ZrxSupportedChainIds = KnownChainIds.EthereumMainnet | KnownChainIds.AvalancheMainnet
